requirements-test.txt
black==19.3b0
coverage==5.2
django-nose==1.4.7
django-test-plus==1.4.0
factory-boy==2.9.2
flake8==3.8.1
isort==5.1.4
mccabe==0.6.1
nose==1.3.7
nose-exclude==0.5.0
nose-progressive==1.5.2
nose-timer==1.0.0
